refactor: `tableGroup`, `table` 패키지 의존성 사이클 제거
